Fix issue handling falsy values
Due to python's handling of falsy values in or statements, an empty dict
or list would be returned as the default value passed to the `get()`
method rather than the intended falsy value.

This fixes it by raising a KeyError rather than falling through in the
`get_key()` method.

Fixes #2.