Add support for Fully-Qualified Puppet Variables (fixes #7)
Puppet (because puppet?) has a pattern of prefixing FQ variables with
two colons, e.g. '::my_puppet_variable'. This sucks for us because folks
who use that formatting in the hiera config (for interpolation) will be
SOL as we use pure python formatting. This commit moves to a regex that
ignores two leading colons at the start of any format variable in the
hiera config.